requirements.txt
Flask==2.0.1
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cpu
torch==1.9.0+cpu
Pillow==7.0.0
gunicorn==20.1.0
numpy==1.20.3
pandas==1.2.4
requests==2.25.1
bs4==0.0.1
